Add scripts for a switch template
A template and script are added in order to run simulations in which it is possible to switch models during the simulation. The record order till the first stopping rule is reached (i.e. n consecutive irrelevant records) is used as prior knowledge for the second model. The output will contain the simulation of:
1. the first model;
2. the second model;
3. a switch from model one to model two.

The basic template was used as building block. The scripts are a work in progress.